# app/docs/swagger.yaml
basePath: /
definitions:
  dtos.CreateShortcutDto:
    properties:
      url:
        type: string
    type: object
  dtos.ErrorDto:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  dtos.ShortcutDto:
    properties:
      id:
        type: string
      short:
        type: string
      url:
        type: string
    type: object
info:
  contact:
    email: luisciber640@gmail.com
    name: Luis Ciber
    url: https://www.luisciber.com/
  description: URL shortener service
  license:
    name: MIT
    url: https://github.com/luicib3r/shortly
  title: Shortly Service
  version: "1.0"
paths:
  /api/shortcut:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ShortcutDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
      summary: Get Shortcuts
      tags:
      - Shortcut
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Shortcut
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateShortcutDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ShortcutDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
      summary: Create Shortcut
      tags:
      - Shortcut
  /api/shortcut/{id}:
    delete:
      parameters:
      - description: Shortcut ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
      summary: Delete Shortcut by ID
      tags:
      - Shortcut
    get:
      parameters:
      - description: Shortcut ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ShortcutDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorDto'
      summary: Get Shortcut by ID
      tags:
      - Shortcut
swagger: "2.0"